Extract shared calculations to RootMeasurePolicy
