+++ compose/ui/ui/src/skikoMain/kotlin/androidx/compose/ui/window/Dialog.skiko.kt
@@ -34,23 +34,12 @@ import androidx.compose.ui.input.pointer.PointerButton
 import androidx.compose.ui.input.pointer.PointerEventType
 import androidx.compose.ui.input.pointer.PointerInputEvent
 import androidx.compose.ui.layout.Layout
-import androidx.compose.ui.layout.MeasurePolicy
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.semantics.dialog
 import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.unit.Constraints
-import androidx.compose.ui.unit.Density
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.IntRect
-import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.center
-import androidx.compose.ui.unit.constrain
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.offset
-import androidx.compose.ui.util.fastForEach
-import androidx.compose.ui.util.fastMap
-import androidx.compose.ui.util.fastMaxBy
-import kotlin.math.min
 
 /**
  * The default scrim opacity.
@@ -164,7 +153,7 @@ private fun DialogLayout(
         val density = LocalDensity.current
         val measurePolicy = rememberDialogMeasurePolicy(
             properties = properties,
-            platformOffset = with(density) { platformOffset }
+            platformOffset = with(density) { platformOffset() }
         ) {
             owner.bounds = it
         }
@@ -181,63 +170,11 @@ private fun rememberDialogMeasurePolicy(
     platformOffset: IntOffset,
     onBoundsChanged: (IntRect) -> Unit
 ) = remember(properties, platformOffset, onBoundsChanged) {
-    MeasurePolicy { measurables, constraints ->
-        val platformConstraints = applyPlatformConstrains(
-            constraints, platformOffset, properties.usePlatformDefaultWidth
-        )
-        val placeables = measurables.fastMap { it.measure(platformConstraints) }
-        val width = placeables.fastMaxBy { it.width }?.width ?: constraints.minWidth
-        val height = placeables.fastMaxBy { it.height }?.height ?: constraints.minHeight
-
-        val placeableSize = IntSize(width, height)
-        val windowSize = IntSize(constraints.maxWidth, constraints.maxHeight)
-        val position = windowSize.center - placeableSize.center
-        onBoundsChanged(IntRect(position, placeableSize))
-
-        layout(windowSize.width, windowSize.height) {
-            placeables.fastForEach {
-                it.place(position.x, position.y)
-            }
-        }
+    RootMeasurePolicy(platformOffset, properties.usePlatformDefaultWidth) { windowSize, contentSize ->
+        val position = windowSize.center - contentSize.center
+        onBoundsChanged(IntRect(position, contentSize))
+        position
     }
-}
-
-internal fun Density.applyPlatformConstrains(
-    constraints: Constraints,
-    platformOffset: IntOffset,
-    usePlatformDefaultWidth: Boolean
-): Constraints {
-    val platformConstraints = constraints.offset(
-        horizontal = -2 * platformOffset.x,
-        vertical = -2 * platformOffset.y
-    )
-    return if (usePlatformDefaultWidth) {
-        platformConstraints.constrain(
-            platformDefaultConstrains(constraints)
-        )
-    } else {
-        platformConstraints
-    }
-}
-
-// TODO: Expect composable getter
-//  Blocked by https://github.com/JetBrains/compose-multiplatform/issues/3373
-internal expect val Density.platformOffset: IntOffset
-
-private fun Density.platformDefaultConstrains(
-    constraints: Constraints
-): Constraints = constraints.copy(
-    maxWidth = min(preferredDialogWidth(constraints), constraints.maxWidth)
-)
-
-// Ported from Android. See https://cs.android.com/search?q=abc_config_prefDialogWidth
-private fun Density.preferredDialogWidth(constraints: Constraints): Int {
-    val smallestWidth = min(constraints.maxWidth, constraints.maxHeight).toDp()
-    return when {
-        smallestWidth >= 600.dp -> 580.dp
-        smallestWidth >= 480.dp -> 440.dp
-        else -> 320.dp
-    }.roundToPx()
 }
 
 private fun PointerInputEvent.isMainAction() =
